requirements.txt
requests
openai
python-dotenv
snowflake-connector-python